backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
openai==1.57.4
chromadb==0.5.23
httpx==0.28.1
python-multipart==0.0.20
beautifulsoup4==4.12.3
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
